refactor(orders): 🪝 create useOrders hook to separate component logic
