+++ src/containers/Orders/Orders.tsx
@@ -1,43 +1,11 @@
-import React, { useState, useEffect, useCallback } from 'react'
-import { getOrders } from '../../services/getOrders'
-import { Order } from '../../types/Orders'
 import { OrderSummary } from '../../components/OrderSummary'
-import { useSession } from '../../context/AuthContext'
 import { handleValidateSuperAdmin } from '../../utils/validateRole'
 import { OrderItem } from '../../components/OrderItem'
-import { useNavigate } from 'react-router-dom'
 import classes from './Orders.module.scss'
+import useOrders from '../../hooks/useOrders'
 
 export const Orders: React.FC = () => {
-  const [orders, setOrders] = useState<Order[]>([])
-  const [loading, setLoading] = useState(true)
-  const [error, setError] = useState<string | null>(null)
-  const { user } = useSession()
-  const navigate = useNavigate()
-
-  useEffect(() => {
-    if (!user) {
-      navigate('/')
-    }
-  }, [user, navigate])
-
-  const fetchOrders = useCallback(async () => {
-    try {
-      setLoading(true)
-      const data = await getOrders()
-      setOrders(data)
-      setError(null)
-    } catch (err) {
-      setError('Failed to fetch orders. Please try again later.')
-      console.error('Failed to fetch orders:', err)
-    } finally {
-      setLoading(false)
-    }
-  }, [])
-
-  useEffect(() => {
-    if (user) fetchOrders()
-  }, [fetchOrders, user])
+  const { error, loading, orders, user } = useOrders()
 
   if (!user) {
     return null
